+++ src/event/index.ts
@@ -1,2 +1,2 @@
 export { EventHandler, EventCallback } from './eventhandler';
-export { Event, TagOpenEvent, TagCloseEvent, AttributeEvent, WhitespaceEvent, DOMReadyEvent } from './event';
+export * from './event';