+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features
 
+- new rule `no-conditional-comments` to disallow usage of conditional comments.
 - handle conditional comments.
 
 ## 0.7.0 (2017-11-04)